model/requirements.txt
pdf2image
streamlit==1.29.0
pillow
transformers
pytesseract